chore(table): split AsyncTable story into multiple files for clarity
Also: . Update snapshots. Add story info.
