+++ src/components/Table/StatefulTable.jsx
@@ -36,9 +36,12 @@ const StatefulTable = ({ data: initialData, expandedData, ...other }) => {
   const [state, dispatch] = useReducer(tableReducer, { data: initialData, view: initialState });
   const isLoading = get(initialState, 'table.loadingState.isLoading');
   // Need to initially sort and filter the tables data
-  useEffect(() => {
-    dispatch(tableRegister({ data: initialData, isLoading }));
-  }, [initialData, isLoading]);
+  useEffect(
+    () => {
+      dispatch(tableRegister({ data: initialData, isLoading }));
+    },
+    [initialData, isLoading]
+  );
 
   const {
     view,
